# bin/gestao-hospitalar-swagger.yml
openapi: 3.0.1
info:
  title: Gestão Hospitalar API
  description: API para gestão de hospital
  contact:
    name: Yuri Nascimento
    url: https://github.com/yuriidiiego
    email: yuriidiiego@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local server
security:
  - bearerAuth: []
...
paths:
...
  /usuarios/login:
    post:
      tags:
        - Usuário
      summary: Loga um usuário
      description: Loga um usuário e retorna um token JWT no cookie
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioLoginRequest"
        required: true
      responses:
        "401":
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Usuário logado com sucesso
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/UsuariorInfoResponse"
  /usuarios/cadastro:
    post:
      tags:
        - Usuário
      summary: Cadastra um usuário
      description: Cadastra um usuário no sistema
      operationId: cadastro
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioCadastroRequest"
        required: true
      responses:
        "200":
          description: Usuário cadastrado com sucesso
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/MensagemResponse"
        "400":
          description: Email ou username já cadastrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statuscode:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        erro:
          type: string
...
    MensagemResponse:
      type: object
      properties:
        mensagem:
          type: string
    UsuarioLoginRequest:
      required:
        - nomeUsuario
        - senha
      type: object
      properties:
        nomeUsuario:
          type: string
          description: Nome de usuário
          example: admin
        senha:
          type: string
          description: Senha do usuário
          example: admin123
    UsuariorInfoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        valueCookie:
          type: string
    UsuarioCadastroRequest:
      required:
        - email
        - nomeUsuario
        - password
      type: object
      properties:
        nomeUsuario:
          maxLength: 20
          minLength: 3
          type: string
          description: Nome de usuário
          example: admin
        email:
          type: string
          description: Email do usuário
          example: admin@example.com
        perfil:
          uniqueItems: true
          type: array
          description: Permissões do usuário
          example:
            - admin
          items:
            type: string
            description: Permissões do usuário
            example: '["admin"]'
        password:
          maxLength: 20
          minLength: 6
          type: string
          description: Senha do usuário
          example: admin123
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT